Rename to services.go. Move lib/auth.Services struct.
